New tests implemented

+++ train-sensor/src/main/java/hu/bme/mit/train/sensor/TrainSensorImpl.java
@@ -22,7 +22,7 @@ public class TrainSensorImpl implements TrainSensor {
 
 	@Override
 	public void overrideSpeedLimit(int speedLimit) {
-		if (speedLimit / controller.getReferenceSpeed() < 0.5 || speedLimit < 0 || speedLimit > 500) {
+		if ((float)speedLimit / (float)controller.getReferenceSpeed() < 0.5 || speedLimit < 0 || speedLimit > 500) {
 			user.setAlarmState(true);
 		} else {
 			user.setAlarmState(false);
